requirements.txt
yfinance==0.2.37
pandas==2.2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlalchemy==2.0.30
dbt-postgres==1.7.14
apache-airflow==2.8.1
streamlit==1.32.2
plotly==5.20.0
